Move revision integration tests to test_program
